# apps/backend/src/main/java/com/travelplanner/api/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Travel Planner API
  version: 1.0.0
  description: |
    Knowledge-grounded, chat-first trip planner.

    **Errors.** Every non-2xx response returns the `ApiErrorResponse` envelope
    `{ code, message, details }`. `code` is a registered `snake_case` identifier; the frontend
    maps it to `common.errors.<code>` and never displays `message` verbatim.

    **Request correlation.** Send `X-Request-Id` (a UUID) on every request. The server echoes
    it on every response — success and failure alike — and puts it in the log MDC. If the
    header is absent or malformed the server generates one.

    **Pagination.** List endpoints take `page`, `page_size` and `sort`, and return the item
    array alongside `page`, `page_size` and `total`.

    **Optimistic concurrency.** Mutations of versioned aggregates carry `expected_version` in
    the request body (ADR 008). A mismatch is `409 version_conflict`; a missing value is
    `400 validation_failed` and is never treated as a force-overwrite.
  license:
    name: UNLICENSED
servers:
  - url: /api/v1
    description: Same-origin path prefix. ADR 006 routes the browser through the Next.js proxy.
...
paths:
...
  /trips/{tripId}/brief:
    get:
      tags: [trips]
      operationId: getTripBrief
      summary: Read the trip brief and its outstanding questions (UC-C1-01, UC-C1-04)
      description: |
        The brief, the trip's status, and the clarification questions the brief still needs — in
        one response, because they are only consistent together. A client that fetched the brief
        and the status separately would, between the two calls, see a `BRIEF_COMPLETE` trip beside
        a brief the agent had just re-opened, and would render a "start research" button the server
        then refuses.

        This is the read that precedes every write, so it is the read that supplies the `version`
        those writes must echo (ADR 008 §1). `version` here is the **brief's**; the trip carries its
        own, and the two advance independently.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TripIdParam'
      responses:
        '200':
          $ref: '#/components/responses/TripBriefState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags: [trips]
      operationId: updateTripBrief
      summary: Save the whole trip brief (UC-C1-01)
      description: |
        The whole-body save the debounced intake form issues (PLAN §4.2).

        **An omitted field clears it.** This is a `PUT` of the whole resource, not a merge:
        `PATCH` is forbidden project-wide (ADR 008 §3), and a `PUT` that read absence as "leave
        unchanged" would give a client no way at all to un-set a field it entered by mistake. To
        change one field without re-sending the rest, use the clarification action below — the
        typed diff ADR 008 §3 exists for exactly that.

        **The response is the new full resource** (ADR 008 §2), including the incremented `version`,
        the recomputed `clarification`, and the trip `status` the save moved the trip to. Status is
        computed from the saved brief on every write, so no request can declare `BRIEF_COMPLETE`
        over a brief that is not.

        **Destination preference is checked against coverage.** A slug the knowledge base has never
        curated is `404 destination_not_covered` carrying `details.supported` (ADR 010 §4) — it is
        never accepted and then ranked near zero, which would read to the user as "we considered it
        and it is a poor match" when the truth is that nobody ever looked. An empty `destinations`
        is a legitimate answer meaning "no preference", and lets C2 rank the whole covered set.

        The status the save can move the trip to is limited to `CLARIFICATION_NEEDED` and
        `BRIEF_COMPLETE`, and a trip that has already left the intake phase is refused with
        `400 validation_failed` rather than dragged back into it.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TripIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTripBriefRequest'
      responses:
        '200':
          $ref: '#/components/responses/TripBriefState'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/TripBriefNotFound'
        '409':
          $ref: '#/components/responses/VersionConflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /trips/{tripId}/brief/actions/answer-clarification:
    post:
      tags: [trips]
      operationId: answerTripBriefClarification
      summary: Answer outstanding clarification questions (UC-C1-04)
      description: |
        Applies typed answers to the fields their questions were asked about, then re-validates —
        the brief either completes or returns a shorter list of questions.

        **A typed `POST` action, not a `PUT`** (ADR 008 §3). PLAN §4.1.3 sketched this as
        `PUT .../brief/clarification`; the ADR supersedes that, and the reason is the one it gives:
        re-sending nine brief fields to answer one question means eight of them were read some time
        ago, and each is a chance to overwrite something the agent changed since — from a form
        control the user never touched.

        Three things are refused rather than ignored, because a silently dropped answer is
        indistinguishable to the user from one the server accepted and then lost:

        - an answer to a question that is not outstanding — it was already answered or never asked,
          and applying it would make this endpoint a second, unvalidated way to write arbitrary
          brief fields;
        - an answer whose value type does not match the question's `type`;
        - two answers to the same question, where "last one wins" would be a coin toss.

        No coverage check: no clarification question sets a destination, so none can name an
        uncovered one.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TripIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerClarificationRequest'
      responses:
        '200':
          $ref: '#/components/responses/TripBriefState'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/VersionConflict'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: tp_session
      description: |
        The session cookie from ADR 002 — `httpOnly`, `SameSite=Lax`, and `Secure` in production.

        Documented as `apiKey in cookie` because OpenAPI 3.0 has no vocabulary for "the browser
        sends this automatically". No client ever sets this header or reads this value: it is
        `httpOnly`, and `lib/api/client.ts` relies on `credentials: 'include'` instead (ADR 006).
  headers:
...
    XRequestId:
      description: |
        Correlation id echoed from the request, or generated when the caller sent none.
        Present on every response, including errors — it is how a user-reported failure is
        found in the logs.
      required: true
      schema:
        type: string
        format: uuid
  parameters:
...
    TripIdParam:
      name: tripId
      in: path
      required: true
      description: |
        The trip's surrogate key. Ownership is *not* carried here — it comes from the session on
        every request, and an id that is not the caller's is answered `404 not_found` rather than
        `403`, so this parameter cannot be used to probe for another user's trips.
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ApiErrorResponse:
      $ref: './errors.yaml#/components/schemas/ApiErrorResponse'
...
    TripStatus:
      type: string
      description: |
        Wizard progress for a trip — the exact vocabulary of the "Trip status" table in
        `USE-CASES.md`, in the order that table lists it (PLAN §3.1).

        **Read-only on the wire.** No request body anywhere on this API accepts a status. It is
        derived from brief completeness, and moved to `ARCHIVED` by the archive action; a client
        able to set it could declare `BRIEF_COMPLETE` over an empty brief and step past the gate in
        front of C2.

        C1 owns `DRAFT`, `CLARIFICATION_NEEDED`, and `BRIEF_COMPLETE`. The rest are published here
        because a client has to be able to render a trip in any of them, not because C1 can set
        them.
      enum:
        - DRAFT
        - BRIEF_COMPLETE
        - CLARIFICATION_NEEDED
        - RESEARCH_QUEUED
        - RESEARCH_RUNNING
        - RESEARCH_READY
        - DESTINATION_SELECTED
        - ITINERARY_READY
        - BOOKING_IN_PROGRESS
        - BOOKED
        - ARCHIVED
      example: DRAFT
...
    Money:
      type: object
      description: |
        An amount and its currency (PLAN §4.0.2-A).

        **`amount` is a decimal string, not a JSON number.** Money is `BigDecimal` in Java and
        `numeric` in Postgres for the whole of its life precisely because binary floating point
        cannot represent `4000.10`; serialising it as a JSON number hands it to a JavaScript client
        as a `double` and undoes that at the last hop. A string survives the round trip exactly.

        The server rejects a negative amount, an unknown currency code, and more decimal places
        than the currency has minor units — 2 for `USD`, 0 for `JPY`.
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: '^\d+(\.\d+)?$'
          description: Non-negative decimal, in the currency's own precision.
          example: '4000.00'
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO 4217.
          example: MYR
    DateRange:
      type: object
      description: |
        An inclusive span of calendar days. `start_date == end_date` is a one-day trip.

        Calendar dates, never instants: "arrive on 3 April" is a fact about the destination, and
        turning it into a point on the UTC timeline needs a timezone the traveller has not chosen
        yet — one that would shift the date across a boundary for anybody east of Greenwich.

        One object rather than two sibling fields on the brief, so a client reads dates as present
        or absent and never has to decide what a start with no end means.
      required: [start_date, end_date]
      properties:
        start_date:
          type: string
          format: date
          example: '2026-04-03'
        end_date:
          type: string
          format: date
          example: '2026-04-12'
    PartySize:
      type: object
      description: |
        How many people are travelling.

        Two counts rather than one total, because the split changes the answer rather than only the
        arithmetic: room occupancy, entry pricing, and the `area_coverage` term of the C2 fit score
        all treat a family of four differently from four adults. At least one adult, and at most 20
        travellers — above that no C4 supplier quotes as a single party, so a brief that exceeded it
        would produce an itinerary that cannot be booked.
      required: [adults, children]
      properties:
        adults:
          type: integer
          format: int32
          minimum: 1
          maximum: 20
          example: 2
        children:
          type: integer
          format: int32
          minimum: 0
          maximum: 19
          example: 0
    DateFlexibility:
      type: string
      description: |
        How far the dates may move (UC-C1-04).

        Kept separate from `dates` rather than modelled as a widened range: a range says which days
        were entered, flexibility says how much C2 may shift them for a cheaper week or a better
        season. Widening the range would lose the original intent and make `seasonality_fit` score
        a window the user never asked for.
      enum: [FIXED, FLEXIBLE_WEEK, FLEXIBLE_MONTH]
      example: FLEXIBLE_WEEK
    TravelPace:
      type: string
      description: |
        How densely the traveller wants their days filled. An ordinal band rather than a number,
        because nobody can answer "how many activities per day" honestly and a number invites a
        precision the itinerary would then be held to.
      enum: [RELAXED, MODERATE, PACKED]
      example: MODERATE
    TravelInterest:
      type: string
      description: |
        What the traveller wants out of the trip — the `interests` half of the C2 fit score
        (PLAN §4.1.2).

        A closed vocabulary, aligned one-to-one with the knowledge base's POI categories. Free text
        would be friendlier to type and useless to rank: "temples and ramen" cannot be intersected
        with anything the corpus stores, so the interest term would silently contribute zero and
        the user would be told their interests were considered when they were not.
      enum:
        - FOOD
        - SIGHTSEEING
        - MUSEUMS
        - NATURE
        - SHOPPING
        - NIGHTLIFE
        - EXPERIENCES
      example: FOOD
    ClarificationType:
      type: string
      description: |
        What kind of value a question is asking for. The frontend picks a control from it, and the
        server picks which field of `ClarificationAnswer` it will read — an answer whose filled
        slot disagrees with this is `400 validation_failed`, never a silently ignored answer.
      enum: [TEXT, NUMBER, MONEY, DATE_RANGE, CHOICE, MULTI_CHOICE]
      example: MONEY
    ClarificationQuestion:
      type: object
      description: |
        One typed question the brief cannot be completed without (UC-C1-04).

        `prompt_key` is a translation key, never a sentence: the frontend resolves it against
        `locales/<lang>/trip_brief.json`, so one payload renders in English and Malay. An English
        string here would be user-facing copy the Malay build could never reach (PLAN §11).
      required: [id, prompt_key, type, options, required]
      properties:
        id:
          type: string
          description: |
            Stable and `snake_case`, and also the name of the brief field the answer sets — it is
            what the answer action routes on. One of `travel_dates`, `date_flexibility`,
            `departure_city`, `budget_max`, `party_size`, `interests`, `pace`.
          example: budget_max
        prompt_key:
          type: string
          description: i18n key under the `trip_brief` namespace.
          example: trip_brief.clarify_budget
        type:
          $ref: '#/components/schemas/ClarificationType'
        options:
          type: array
          description: |
            The permitted values for `CHOICE` and `MULTI_CHOICE`, and empty for every other type.
            Carried in the question rather than fetched separately, so a client renders the control
            from one payload.
          items:
            type: string
          example: []
        required:
          type: boolean
          description: |
            Whether the brief can complete without an answer. Every question C1 generates is
            required, because it only ever asks about a field completeness depends on; the flag is
            here for the LLM-authored questions task 19 adds, which can be optional.
          example: true
    Clarification:
      type: object
      description: |
        The typed gap between this brief and a complete one (UC-C1-04, PLAN §4.1.3).

        **Derived on every read and every write, never stored.** The questions are a pure function
        of the brief's current fields, so the completeness rule and the question list are one
        computation read two ways. A stored question table would be a second copy of that rule, and
        the two would disagree the first time the agent wrote a field through `update_trip_brief`
        without passing through the endpoint that refreshed the table.

        Always present, and empty exactly when the brief is complete — so a client never has to
        tell "no questions" apart from "questions not computed".

        Destination preference is deliberately never asked about: an empty preference is a
        legitimate answer that lets C2 rank the whole covered set, and demanding one would make the
        product unusable for the traveller whose actual question is "where should I go?".
      required: [questions]
      properties:
        questions:
          type: array
          description: In a stable order, so answering one does not reshuffle the form.
          items:
            $ref: '#/components/schemas/ClarificationQuestion'
    ClarificationAnswer:
      type: object
      description: |
        One answer to one outstanding question — the typed diff ADR 008 §3 requires of an action
        body.

        **Exactly one slot must be filled**, and it must be the one the question's `type` names.
        The alternative — a single `value` string parsed according to the type — cannot carry a
        budget (an amount plus a currency) or a date range (two dates) without inventing a text
        encoding, and a text encoding is a parser that fails at runtime in a place this contract
        cannot describe.
      required: [question_id]
      properties:
        question_id:
          type: string
          description: The `ClarificationQuestion.id` being answered.
          example: budget_max
        text:
          type: string
          nullable: true
          description: For `TEXT`.
        number:
          type: integer
          format: int32
          nullable: true
          description: For `NUMBER`.
        money:
          allOf:
            - $ref: '#/components/schemas/Money'
          nullable: true
          description: For `MONEY`.
        date_range:
          allOf:
            - $ref: '#/components/schemas/DateRange'
          nullable: true
          description: For `DATE_RANGE`.
        choice:
          type: string
          nullable: true
          description: For `CHOICE`. One of the question's `options`.
        choices:
          type: array
          nullable: true
          description: |
            For `MULTI_CHOICE`. An empty array counts as unfilled — selecting nothing is not an
            answer, it is the question restated.
          items:
            type: string
    TripBrief:
      type: object
      description: |
        The structured requirement set for one trip — what C1 produces and C2 consumes — together
        with the trip status it implies and the questions it still needs.

        The three travel in one payload because they are only consistent together, and this is the
        body of every brief read **and** of every successful brief mutation (ADR 008 §2: "every
        mutation returns the new full resource with its incremented version"). After a save the
        client holds exactly what the server holds, so the debounced form's next keystroke does not
        race a refetch it would otherwise have to make.

        Every field is nullable or empty, because a brief under construction is incomplete by
        definition and PLAN §4.1.3 answers incompleteness with `clarification` rather than a
        rejected save. What is *present* is always valid.

        There is no brief id: the brief is the trip's one child, addressed only through `tripId`.
      required:
        - trip_id
        - status
        - destinations
        - interests
        - clarification
        - version
        - created_at
        - updated_at
      properties:
        trip_id:
          type: string
          format: uuid
        status:
          allOf:
            - $ref: '#/components/schemas/TripStatus'
          description: |
            The **trip's** status, carried here because it changes as a consequence of saving the
            brief. A client that had to fetch it separately would, in the gap, render a "start
            research" button for a brief the agent had just re-opened.
        destinations:
          type: array
          maxItems: 10
          description: |
            Candidate `destination.slug` values, most-preferred first. Empty means "no preference",
            which is a legitimate answer rather than a gap. Every slug is checked against the
            covered set on save (ADR 010 §4).
          items:
            type: string
          example: [kuala-lumpur]
        dates:
          allOf:
            - $ref: '#/components/schemas/DateRange'
          nullable: true
        date_flexibility:
          allOf:
            - $ref: '#/components/schemas/DateFlexibility'
          nullable: true
        departure_city:
          type: string
          maxLength: 120
          nullable: true
          description: |
            Free text, deliberately not a slug. People depart from places the knowledge base has
            never curated, and requiring coverage on the origin would refuse a valid trip.
          example: Kuala Lumpur
        budget:
          allOf:
            - $ref: '#/components/schemas/Money'
          nullable: true
          description: The ceiling for the whole trip, not a per-day figure.
        party:
          allOf:
            - $ref: '#/components/schemas/PartySize'
          nullable: true
        interests:
          type: array
          maxItems: 16
          description: Duplicates are collapsed; order is not meaningful.
          items:
            $ref: '#/components/schemas/TravelInterest'
        pace:
          allOf:
            - $ref: '#/components/schemas/TravelPace'
          nullable: true
        clarification:
          $ref: '#/components/schemas/Clarification'
        version:
          type: integer
          format: int32
          minimum: 0
          description: |
            The **brief's** optimistic lock, and what `expected_version` means on both brief write
            paths. The trip carries its own; the two advance independently (ADR 008 §1).
          example: 7
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UpdateTripBriefRequest:
      allOf:
        - $ref: '#/components/schemas/ExpectedVersion'
        - type: object
          description: |
            A `PUT` of the whole resource: **an omitted field clears it.** `PATCH` is forbidden
            project-wide (ADR 008 §3), and a `PUT` that read absence as "leave unchanged" would
            give a client no way at all to un-set a field it entered by mistake.
          properties:
            destinations:
              type: array
              maxItems: 10
              items:
                type: string
            dates:
              allOf:
                - $ref: '#/components/schemas/DateRange'
              nullable: true
            date_flexibility:
              allOf:
                - $ref: '#/components/schemas/DateFlexibility'
              nullable: true
            departure_city:
              type: string
              maxLength: 120
              nullable: true
            budget:
              allOf:
                - $ref: '#/components/schemas/Money'
              nullable: true
            party:
              allOf:
                - $ref: '#/components/schemas/PartySize'
              nullable: true
            interests:
              type: array
              maxItems: 16
              items:
                $ref: '#/components/schemas/TravelInterest'
            pace:
              allOf:
                - $ref: '#/components/schemas/TravelPace'
              nullable: true
    AnswerClarificationRequest:
      allOf:
        - $ref: '#/components/schemas/ExpectedVersion'
        - type: object
          description: |
            At least one answer. An action that answers nothing would consume a version and change
            no state, leaving the client unable to tell which of the two happened.
          required: [answers]
          properties:
            answers:
              type: array
              minItems: 1
              items:
                $ref: '#/components/schemas/ClarificationAnswer'
    ExpectedVersion:
      type: object
      description: |
        Optimistic-concurrency fragment (ADR 008 §2). Every mutation of a versioned aggregate
        composes this into its request body:

        ```yaml
        UpdateTripBriefRequest:
          allOf:
            - $ref: '#/components/schemas/ExpectedVersion'
            - type: object
              properties: { destinations: { ... } }
        ```

        The version travels in the body rather than in `ETag`/`If-Match` precisely because a
        body field is expressible in OpenAPI and flows through codegen into the typed client;
        header-based concurrency would need hand-plumbed handling in the generated-client
        layer this project forbids editing.

        Omitting `expected_version` is `400 validation_failed` — **never** a force-overwrite.
        A mismatch is `409 version_conflict` carrying `details.current_version`.
      required: [expected_version]
      properties:
        expected_version:
          type: integer
          format: int32
          minimum: 0
          description: The version the client based its edit on.
          example: 7
  responses:
    ValidationFailed:
      description: |
        The request failed schema or constraint validation. `details.fields` maps field name to
        an English message (`ValidationFailedDetails`).
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: validation_failed
            message: The request is not valid.
            details:
              fields:
                expected_version: must not be null
    Unauthorized:
      description: No valid session. The caller must sign in.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: unauthorized
            message: Authentication is required.
            details: {}
...
    NotFound:
      description: |
        No resource at this path, or none owned by the caller. Ownership failures return 404
        rather than 403 so the API does not confirm that another user's resource exists.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: not_found
            message: The requested resource does not exist.
            details: {}
...
    TripBriefState:
      description: |
        The brief, the trip status it implies, and the outstanding questions — one consistent
        payload, and the **new full resource** after any mutation (ADR 008 §2). `version` is the
        brief's, and is what `expected_version` means on both brief write paths.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TripBrief'
    TripBriefNotFound:
      description: |
        Nothing to write, for one of two reasons that share `404`.

        - `not_found` — no trip with this id, **or** one that is not the caller's. Deliberately
          identical, so the endpoint cannot confirm that another user's trip exists.
        - `destination_not_covered` — a named destination has no curated knowledge (ADR 010 §4).
          `details.requested` echoes the slug so a typo is visible and `details.supported` lists the
          ones that are covered, because the honest answer to "can you plan Osaka?" is "no, but
          here is what we do know". It is never accepted and then ranked near zero, which would
          read as "we considered it and it is a poor match" when nobody ever looked.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: destination_not_covered
            message: No curated travel knowledge exists for 'osaka' yet.
            details:
              requested: osaka
              supported: [kuala-lumpur, penang, singapore]
    VersionConflict:
      description: |
        The `expected_version` in the request body does not match the stored version
        (ADR 008). `details.current_version` carries the server's version so the client can
        re-read, re-apply its edit, and retry. There is no implicit winner.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: version_conflict
            message: The resource was modified by someone else.
            details:
              current_version: 9
...
    InternalError:
      description: Unhandled server fault. `details` is always empty — diagnostics stay in the logs.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: internal_error
            message: An unexpected error occurred.
            details: {}